refactor: move DashboardCardInnerErrorView to its own file
